Reorganized files and fixed error message
